FIX ArticleImageRepository : Id/Long

+++ src/main/java/com/forge_miniatures/repository/ArticleImageRepository.java
@@ -3,5 +3,5 @@ package com.forge_miniatures.repository;
 import com.forge_miniatures.entity.ArticleImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ArticleImageRepository extends JpaRepository<ArticleImage,Integer> {
+public interface ArticleImageRepository extends JpaRepository<ArticleImage,Long> {
 }
